create debian package

+++ com.erichsteiger.eopt.bo/src/main/java/com/erichsteiger/eopt/bo/slideshow/SlideFooter.java
@@ -65,7 +65,7 @@ public class SlideFooter {
     this.footerRight = footerRight;
   }
 
-  public double getFontSize() {
+  public int getFontSize() {
     return fontSize;
   }
